% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annofuns_GO}
\alias{annofuns_GO}
\title{GO annotations to pathways object}
\format{GO annotations to pathways object}
\usage{
data(annofuns_GO)
}
\value{
GO annotations to pathways object
}
\description{
GO annotations to pathways object
}
\keyword{datasets}
