% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annofuns_uniprot}
\alias{annofuns_uniprot}
\title{Uniprot annotations to pathways object}
\format{Uniprot annotations to pathways object}
\usage{
data(annofuns_uniprot)
}
\value{
Uniprot annotations to pathways object
}
\description{
Uniprot annotations to pathways object
}
\keyword{datasets}
