% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{entrez_hgnc}
\alias{entrez_hgnc}
\title{Entrez - HGNC annotations object}
\format{Entrez - HGNC annotations object}
\usage{
data(entrez_hgnc)
}
\value{
Entrez - HGNC annotations object
}
\description{
Entrez - HGNC annotations object
}
\keyword{datasets}
