% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{go_bp_annots}
\alias{go_bp_annots}
\title{GO annotations object}
\format{GO annotations object}
\usage{
data(go_bp_annots)
}
\value{
GO annotations object
}
\description{
GO annotations object
}
\keyword{datasets}
