% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{go_bp_frame}
\alias{go_bp_frame}
\title{GO graph information}
\format{GO graph information}
\usage{
data(go_bp_frame)
}
\value{
GO graph information
}
\description{
GO graph information
}
\keyword{datasets}
