% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load.pseudo.mgi}
\alias{load.pseudo.mgi}
\title{Loads object with pseudo graph information}
\usage{
load.pseudo.mgi(species, group.by)
}
\arguments{
\item{species}{Species of the samples.}
}
\value{
Pseudo graph information object
}
\description{
Loads object with pseudo graph information
}
\examples{
load.pseudo("hsa", "uniprot")

}
