% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta_graph_info_hsa}
\alias{meta_graph_info_hsa}
\title{Object of pathways information, HSA}
\format{Object of pathways information, HSA}
\usage{
data(meta_graph_info_hsa)
}
\value{
Object of pathways information, HSA
}
\description{
Object of pathways information, HSA
}
\keyword{datasets}
