% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta_graph_info_mmu}
\alias{meta_graph_info_mmu}
\title{Object of pathways information, MMU}
\format{Object of pathways information, MMU}
\usage{
data(meta_graph_info_mmu)
}
\value{
Object of pathways information, MMU
}
\description{
Object of pathways information, MMU
}
\keyword{datasets}
