% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta_graph_info_rno}
\alias{meta_graph_info_rno}
\title{Object of pathways information, RNO}
\format{Object of pathways information, RNO}
\usage{
data(meta_graph_info_rno)
}
\value{
Object of pathways information, RNO
}
\description{
Object of pathways information, RNO
}
\keyword{datasets}
