% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pmgi_hsa_genes}
\alias{pmgi_hsa_genes}
\title{Object of pathways information grouped by genes, HSA}
\format{Object of pathways information grouped by genes, HSA}
\usage{
data(pmgi_hsa_genes)
}
\value{
Object of pathways information grouped by genes, HSA
}
\description{
Object of pathways information grouped by genes, HSA
}
\keyword{datasets}
