% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pmgi_hsa_GO}
\alias{pmgi_hsa_GO}
\title{Object of pathways information grouped by GO functions, HSA}
\format{Object of pathways information grouped by GO functions, HSA}
\usage{
data(pmgi_hsa_GO)
}
\value{
Object of pathways information grouped by GO functions, HSA
}
\description{
Object of pathways information grouped by GO functions, HSA
}
\keyword{datasets}
