% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pmgi_hsa_uniprot}
\alias{pmgi_hsa_uniprot}
\title{Object of pathways information grouped by Uniprot functions, HSA}
\format{Object of pathways information grouped by Uniprot functions, HSA}
\usage{
data(pmgi_hsa_uniprot)
}
\value{
Object of pathways information grouped by Uniprot functions, HSA
}
\description{
Object of pathways information grouped by Uniprot functions, HSA
}
\keyword{datasets}
