% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pmgi_mmu_genes}
\alias{pmgi_mmu_genes}
\title{Object of pathways information grouped by genes, MMU}
\format{Object of pathways information grouped by genes, MMU}
\usage{
data(pmgi_mmu_genes)
}
\value{
Object of pathways information grouped by genes, MMU
}
\description{
Object of pathways information grouped by genes, MMU
}
\keyword{datasets}
