% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pmgi_mmu_GO}
\alias{pmgi_mmu_GO}
\title{Object of pathways information grouped by GO functions, MMU}
\format{Object of pathways information grouped by GO functions, MMU}
\usage{
data(pmgi_mmu_GO)
}
\value{
Object of pathways information grouped by GO functions, MMU
}
\description{
Object of pathways information grouped by GO functions, MMU
}
\keyword{datasets}
