% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pmgi_rno_genes}
\alias{pmgi_rno_genes}
\title{Object of pathways information grouped by genes, RNO}
\format{Object of pathways information grouped by genes, RNO}
\usage{
data(pmgi_rno_genes)
}
\value{
Object of pathways information grouped by genes, RNO
}
\description{
Object of pathways information grouped by genes, RNO
}
\keyword{datasets}
