% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pmgi_rno_GO}
\alias{pmgi_rno_GO}
\title{Object of pathways information grouped by GO functions, RNO}
\format{Object of pathways information grouped by GO functions, RNO}
\usage{
data(pmgi_rno_GO)
}
\value{
Object of pathways information grouped by GO functions, RNO
}
\description{
Object of pathways information grouped by GO functions, RNO
}
\keyword{datasets}
