% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pmgi_rno_uniprot}
\alias{pmgi_rno_uniprot}
\title{Object of pathways information grouped by Uniprot functions, RNO}
\format{Object of pathways information grouped by Uniprot functions, RNO}
\usage{
data(pmgi_rno_uniprot)
}
\value{
Object of pathways information grouped by Uniprot functions, RNO
}
\description{
Object of pathways information grouped by Uniprot functions, RNO
}
\keyword{datasets}
