% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uni_bp_annots}
\alias{uni_bp_annots}
\title{Uniprot annotations object}
\format{Uniprot annotations object}
\usage{
data(uni_bp_annots)
}
\value{
Uniprot annotations object
}
\description{
Uniprot annotations object
}
\keyword{datasets}
